// libs/aggregation/aggregation-adapters/src/lib/adapters/partners/budgetinsights/models/BudgetInsightConnectionState.ts
export enum BudgetInsightConnectionState {
  SCA_REQUIRED = 'SCARequired',
  DECOUPLED = 'decoupled',
  ADDITIONNAL_INFORMATION = 'additionalInformationNeeded',
  ACTION_NEEDED = 'actionNeeded',
  PASSWORD_EXPIRED = 'passwordExpired',
  BUG = 'bug',
  WEBSITE_UNAVAILABLE = 'websiteUnvailable',
  WRONG_PASSWORD = 'wrongpass',
}